+++ CHANGES.rst
@@ -39,6 +39,9 @@ Unreleased
 -   Fix setting CSP header options on the response. :pr:`2237`
 -   Fix an issue with with the interactive debugger where lines would
     not expand on click for very long tracebacks. :pr:`2239`
+-   The interactive debugger handles displaying an exception that does
+    not have a traceback, such as from ``ProcessPoolExecutor``.
+    :issue:`2217`
 
 
 Version 2.0.1